Version: 2006-09-12 ksh93s-
Sourced by using the relevant ksh93-integration diff:
4faadb3b298232d391fd8c9ad38e113ad034b3f025a1610039c00d2568841bfb  2006-09-12-reverse.diff

+++ src/lib/libast/sfio/sfpoll.c
@@ -231,7 +231,7 @@ int		tm;	/* time in millisecs for select/poll	*/
 
 		/* announce status */
 		if(f->disc && f->disc->exceptf)
-			(*f->disc->exceptf)(f,SF_READY,(Void_t*)status[c],f->disc);
+			(*f->disc->exceptf)(f,SF_READY,(Void_t*)(long)status[c],f->disc);
 
 		if(c > r) /* move to front of list */
 		{	fa[c] = fa[r];
